+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Changelog
 (most recent first)
 
+### v0.1 2018-03-10:
+* Added ancillary, very alpha, **`thmigpoll.py`** - a rudimentary script to list respondents to a poll; entirely dependent on an RSS feed of a pnut.io hashtag.  
+* Edited `thmigpen` pnut channel code comments.
+
 ### v0.0 2018-03-07:
 * First commit with basic functionality
  * Builds a list of only 1 year of first Mondays before the \#ThemeMonday second Mondays, starting 2018-01-01.